pipeline/requirements.programma.txt
requests==2.32.3
html==1.16
pytest-snapshot==0.9.0
unittest==0.0
jsonschema==4.23.0

